// backend/src/common/enums.ts
export enum Role {
  EMPLOYEE = 'EMPLOYEE',
  MANAGER = 'MANAGER',
  ADMIN = 'ADMIN'
}

export enum GoalStatus {
  DRAFT = 'DRAFT',
  SUBMITTED = 'SUBMITTED',
  UNDER_REVIEW = 'UNDER_REVIEW',
  APPROVED = 'APPROVED',
  REWORK_REQUESTED = 'REWORK_REQUESTED',
  LOCKED = 'LOCKED',
  COMPLETED = 'COMPLETED'
}

export enum CheckInStatus {
  NOT_STARTED = 'NOT_STARTED',
  ON_TRACK = 'ON_TRACK',
  COMPLETED = 'COMPLETED',
  DELAYED = 'DELAYED'
}